faz2(2.5): entity MaxLength + Car index'leri + MinPrice/MaxPrice long + gorsel magic-byte
- Car/ChronicIssue/ProsCons/MileageMilestone kisa alanlarina [MaxLength]
  (uzun serbest metinler text kaliyor - prod veri kaybi riski yok)
- Car.MinPrice/MaxPrice: int -> long (entity + DTO'lar)
- Car.Brand/Segment/ReliabilityScore uzerine index (filtre/siralama)
- AdminCarController: gorsel yuklemede magic-byte (dosya imzasi) kontrolu
- Migration: Faz2DataIntegrity (artimli, DropTable/DropColumn yok)

+++ OtoRehber.Domain/DTOs/CarCreateDto.cs
@@ -28,11 +28,11 @@ namespace OtoRehber.Domain.DTOs
 
         [Required]
         [Range(100000, 50000000)]
-        public int MinPrice { get; set; }
+        public long MinPrice { get; set; }
 
         [Required]
         [Range(100000, 50000000)]
-        public int MaxPrice { get; set; }
+        public long MaxPrice { get; set; }
 
         [Range(0, 10000, ErrorMessage = "Geçerli bir bakım maliyeti girin.")]
         public int EstimatedMaintenanceCostEUR { get; set; }
